Adding new package to split the news queries
