+++ news_queries/app_init_news.py
