+++ news_scraping/berlingske.py
@@ -1,7 +1,7 @@
 import bs4 as bs
 import datetime
 import requests
-from data_queries.news_saving import news_save
+from news_queries.news_saving import news_save
 
 news_saved = 0
 news_found = 0
